Setup for implementing NPC death
