+++ room/room.go
@@ -25,16 +25,6 @@ type Exit struct {
 	Room   interfaces.RoomI
 }
 
-type NPC struct {
-	Id        int    `json:"id"`
-	Name      string `json:"name"`
-	Desc      string `json:"description"`
-	MaxHealth int    `json:"max_health"`
-	Health    int    `json:"health"`
-	Str       int    `json:"strength"`
-	End       int    `json:"endurance"`
-}
-
 type RoomFinder struct {
 	roomMap map[int]int // Room ID to index in []Room
 	Rooms   []*Room
@@ -159,32 +149,4 @@ func (e *Exit) GetKey() string {
 
 func (e *Exit) GetDesc() string {
 	return e.Desc
-}
-
-func (n *NPC) GetName() string {
-	return n.Name
-}
-
-func (n *NPC) GetDesc() string {
-	return n.Desc
-}
-
-func (n *NPC) GetHealth() int {
-	return n.Health
-}
-
-func (n *NPC) GetMaxHealth() int {
-	return n.MaxHealth
-}
-
-func (n *NPC) GetEnd() int {
-	return n.End
-}
-
-func (n *NPC) GetStr() int {
-	return n.Str
-}
-
-func (n *NPC) SetHealth(newHealth int) {
-	n.Health = newHealth
 }
